Simple "search" based on pandas
Really its not searching, its slicing a pandas data frame based on the
database load as a CSV file.  It can slice on title, artist or both,
and then outputs the results to a CSV file.  requirements are
documented in header of pandasearch.py, I think I hit them all.

test.csv is just the database that was in master, it has been resaved
as a CSV file, but nothing else has changed.